[BE] fix : 트랜잭션 내부 호출 오류 수정
