+++ backend/src/main/java/moim_today/application/admin/user_inquiry/UserInquiryService.java
@@ -5,11 +5,10 @@ import moim_today.dto.admin.user_inquiry.UserInquiryAnswerRequest;
 import moim_today.dto.admin.user_inquiry.UserInquiryRespondRequest;
 import moim_today.dto.admin.user_inquiry.UserInquiryResponse;
 import moim_today.dto.mail.MailSendRequest;
+import moim_today.implement.admin.user_inquiry.UserInquiryFinder;
 import moim_today.implement.member.MemberFinder;
 import moim_today.persistence.entity.admin.UserInquiryJpaEntity;
 import moim_today.persistence.repository.admin.user_inquiry.UserInquiryRepository;
-import moim_today.persistence.repository.department.department.DepartmentRepository;
-import moim_today.persistence.repository.member.MemberRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -22,37 +21,27 @@ public class UserInquiryService {
 
     private final MailService mailService;
     private final MemberFinder memberFinder;
-    private final DepartmentRepository departmentRepository;
+    private final UserInquiryFinder userInquiryFinder;
     private final UserInquiryRepository userInquiryRepository;
 
     public UserInquiryService(final MailService mailService,
                               final MemberFinder memberFinder,
-                              final DepartmentRepository departmentRepository,
+                              final UserInquiryFinder userInquiryFinder,
                               final UserInquiryRepository userInquiryRepository) {
         this.mailService = mailService;
         this.memberFinder = memberFinder;
-        this.departmentRepository = departmentRepository;
+        this.userInquiryFinder = userInquiryFinder;
         this.userInquiryRepository = userInquiryRepository;
     }
 
     @Transactional(readOnly = true)
     public List<UserInquiryResponse> getAllUserInquiry(final long universityId) {
-        return userInquiryRepository.getAllUserInquiryByUniversityId(universityId).stream()
-                .map(userInquiryJpaEntity -> {
-                    String departmentName = departmentRepository.getById(userInquiryJpaEntity.getDepartmentId()).getDepartmentName();
-                    return UserInquiryResponse.of(userInquiryJpaEntity, departmentName);
-                })
-                .toList();
+        return userInquiryFinder.getAllUserInquiry(universityId);
     }
 
     @Transactional(readOnly = true)
     public List<UserInquiryResponse> getAllNotAnsweredUserInquiry(final long universityId) {
-        return userInquiryRepository.getAllNotAnsweredUserInquiry(universityId).stream()
-                .map(userInquiryJpaEntity -> {
-                    String departmentName = departmentRepository.getById(userInquiryJpaEntity.getDepartmentId()).getDepartmentName();
-                    return UserInquiryResponse.of(userInquiryJpaEntity, departmentName);
-                })
-                .toList();
+        return userInquiryFinder.getAllNotAnsweredUserInquiry(universityId);
     }
 
     @Transactional
@@ -62,19 +51,9 @@ public class UserInquiryService {
     }
 
     public void respondUserInquiry(final UserInquiryRespondRequest userInquiryRespondRequest) {
-        String subject = makeResponseSubject(userInquiryRespondRequest.userInquiryId());
-        String userEmail = getUserEmail(userInquiryRespondRequest.memberId());
+        String subject = userInquiryFinder.makeResponseSubject(userInquiryRespondRequest.userInquiryId());
+        String userEmail = memberFinder.getMemberProfile(userInquiryRespondRequest.memberId()).email();
         MailSendRequest mailSendRequest = userInquiryRespondRequest.toMailSendRequest(subject, userEmail);
         mailService.sendUserInquiryResponseMail(mailSendRequest, userInquiryRespondRequest.responseContent());
-    }
-
-    public String getUserEmail(final long memberId){
-        return memberFinder.getMemberProfile(memberId).email();
-    }
-
-    @Transactional(readOnly = true)
-    public String makeResponseSubject(final long userInquiryId) {
-        UserInquiryJpaEntity findUserInquiry = userInquiryRepository.getById(userInquiryId);
-        return USER_INQUIRY_RESPONSE_SUBJECT_PREFIX.value() + findUserInquiry.getInquiryTitle();
     }
 }
